Fix circular dependencies (#5651)
* Remove circular dependencies in project form

* Remove lazy loading of NotFound since it is statically imported elsewhere anyway

* Remove circular dependency for empty person object

* Avoid circular dependency for user form
