+++ src/pages/basic_data/institution_admin/edit_user/UserFormDialog.tsx
@@ -14,7 +14,6 @@ import { useMutation, useQuery } from '@tanstack/react-query';
 import { Form, Formik, FormikProps } from 'formik';
 import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
-import * as Yup from 'yup';
 import { fetchProtectedResource } from '../../../../api/commonApi';
 import { updateCristinPerson } from '../../../../api/cristinApi';
 import { createUser, fetchUser, updateUser } from '../../../../api/roleApi';
@@ -24,31 +23,16 @@ import { RootState } from '../../../../redux/store';
 import { CristinPerson, Employment, InstitutionUser, RoleName, UserRole } from '../../../../types/user.types';
 import { getIdentifierFromId } from '../../../../utils/general-helpers';
 import { getUsername, getValueByKey } from '../../../../utils/user-helpers';
-import { personDataValidationSchema } from '../../../../utils/validation/basic_data/addEmployeeValidation';
 import { AffiliationFormSection } from './AffiliationFormSection';
 import { PersonFormSection } from './PersonFormSection';
 import { RolesFormSection } from './RolesFormSection';
 import { TasksFormSection, rolesWithAreaOfResponsibility } from './TasksFormSection';
-
-export enum UserFormFieldName {
-  Employments = 'person.employments',
-  Roles = 'user.roles',
-  ViewingScope = 'user.viewingScope.includedUnits',
-}
-
-const validationSchema = Yup.object().shape({
-  person: personDataValidationSchema,
-});
+import { UserFormData, UserFormFieldName, validationSchema } from './userFormHelpers';
 
 interface UserFormDialogProps extends Pick<DialogProps, 'open'> {
   existingPerson: CristinPerson | string;
   existingUser?: InstitutionUser;
   onClose: () => void;
-}
-
-export interface UserFormData {
-  person?: CristinPerson;
-  user?: InstitutionUser;
 }
 
 export const UserFormDialog = ({ open, onClose, existingUser, existingPerson }: UserFormDialogProps) => {
